refactor(components): split core roles
