+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Shared/Components/CoreComponent.Init.cs
@@ -1,0 +1,17 @@
+using Sandbox.ModAPI;
+
+namespace ShipCoreFramework
+{
+    internal partial class CoreComponent
+    {
+        internal bool Init(IMyFunctionalBlock coreBlock, GridComponent gridComponent, GroupComponent groupComponent)
+        {
+            CoreBlock = coreBlock;
+            GridComponent = gridComponent;
+            _groupComponent = groupComponent;
+            SubtypeId = CoreBlock.BlockDefinition.SubtypeId;
+
+            return Session.IsServer ? InitAuthoritative() : InitClientObserver();
+        }
+    }
+}
